+++ SharePointFramework/ProjectExtensions/src/riskAction/components/RiskAction/NewRiskActionPanel/ResponsibleField/ResponsibleField.tsx
@@ -1,0 +1,52 @@
+import { Combobox, Field, Option, Persona } from '@fluentui/react-components'
+import React, { FC, useContext, useEffect, useState } from 'react'
+import { RiskActionFieldCustomizerContext } from '../../../../context'
+import { IResponsibleFieldProps } from './types'
+
+export const ResponsibleField: FC<IResponsibleFieldProps> = (props) => {
+    const context = useContext(RiskActionFieldCustomizerContext)
+    const [value, setValue] = useState('')
+    const [matchingUsers, setMatchingUsers] = useState([])
+
+    useEffect(() => {
+        context.dataAdapter
+            .clientPeoplePickerSearchUser(value, [])
+            .then((users) => setMatchingUsers(users))
+    }, [value])
+
+    return (
+        <Field label='Ansvarlig'>
+            <Combobox
+                onOptionSelect={(_e, data) => {
+                    props.onChange(data.optionValue)
+                }}
+                onChange={(e) => {
+                    setValue(e.target.value)
+                }}
+            >
+                {matchingUsers.map((user, index) => (
+                    <Option key={index} text={user.text} value={user.secondaryText}>
+                        <Persona
+                            avatar={{ color: 'colorful', 'aria-hidden': true }}
+                            name={user.text}
+                            presence={{
+                                status: 'unknown'
+                            }}
+                            secondaryText={user.secondaryText}
+                        />
+                    </Option>
+                ))}
+                {matchingUsers.length === 0 ? (
+                    <Option key='no-results' text=''>
+                        Ingen brukere funnet.
+                    </Option>
+                ) : null}
+            </Combobox>
+        </Field>
+    )
+}
+
+ResponsibleField.displayName = 'ResponsibleField'
+ResponsibleField.defaultProps = {
+    onChange: () => {}
+}
