+++ SharePointFramework/ProjectExtensions/src/riskAction/components/RiskAction/NewRiskActionPanel/ResponsibleField/index.ts
@@ -1,0 +1,1 @@
+export * from './ResponsibleField'
