+++ SharePointFramework/ProjectExtensions/src/riskAction/components/RiskAction/NewRiskActionPanel/ResponsibleField/types.ts
@@ -1,0 +1,3 @@
+export interface IResponsibleFieldProps {
+    onChange?: (value: string) => void
+}
