--- SharePointFramework/ProjectExtensions/src/riskAction/components/RiskAction/NewRiskActionPanel/ResposibleField/index.ts
@@ -1,1 +1,0 @@
-export * from './ResposibleField'
